#3668 - Disable tutorial for now
- Moved tutorial code to separate module that is not included in the final JAR
- If for some reason the JAR is in the classpath though, the tutorial will be enabled
