+++ inception/inception-tutorial/marker-wicket-module
@@ -1,0 +1,1 @@
+Marker file which activates the profile "wicket-module" from the parent POM.
