+++ inception/inception-tutorial/src/main/java/de/tudarmstadt/ukp/inception/tutorial/enjoyhint.js
