+++ inception/inception-tutorial/src/main/java/de/tudarmstadt/ukp/inception/tutorial/tutorial.js
